src/windows_requirements.txt
amqp==5.1.1
aniso8601==9.0.1
asgiref==3.5.0
asttokens==2.2.1
backcall==0.2.0
billiard==3.6.4.0
bleach==6.0.0
boto3==1.26.127
botocore==1.29.127
Brotli==1.0.9
celery==5.2.7
certifi==2022.6.15.1
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
configloader==1.0.1
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.2.35
cryptography==37.0.4
cssselect2==0.7.0
decorator==5.1.1
distlib==0.3.6
Django==4.0.3
django-appconf==1.0.5
django-celery-beat==2.5.0
django-celery-results==2.5.0
django-chunk-upload-handlers==0.0.13
django-compressor==4.1
django-cors-headers==3.11.0
django-crispy-forms==1.14.0
django-dotenv==1.4.2
django-extensions==3.2.1
django-filter==23.1
django-model-utils==4.3.1
django-mysql==4.7.1
django-otp==1.1.4
django-post-office==3.6.3
django-storages==1.13.2
django-timezone-field==5.0
djangorestframework==3.13.1
djangorestframework-camel-case==1.3.0
djangorestframework-recursive==0.1.2
djangorestframework-simplejwt==5.1.0
docopt==0.6.2
drf-flex-fields==1.0.0
drf-writable-nested==0.7.0
drf-yasg==1.21.4
et-xmlfile==1.1.0
executing==1.2.0
filelock==3.9.0
Flask==2.2.2
Flask-RESTful==0.3.9
fonttools==4.40.0
geographiclib==1.52
geopy==2.2.0
gunicorn==20.1.0
html5lib==1.1
identify==2.5.17
idna==3.3
ifsc-expand==1.0.0
ifscApi==0.2
importlib-metadata==5.0.0
inflection==0.5.1
ipython==8.13.0
itsdangerous==2.1.2
itypes==1.2.0
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
jsonfield==3.1.0
kombu==5.2.4
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
nested-multipart-parser==1.5.0
nodeenv==1.7.0
num2words==0.5.12
numpy==1.23.4
openpyxl==3.0.10
packaging==23.0
pandas==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.6.2
pre-commit==2.21.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.6
pure-eval==0.2.2
pycparser==2.21
pydyf==0.6.0
Pygments==2.14.0
PyJWT==2.3.0
PyMySQL==1.0.2
pyphen==0.14.0
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
qrcode==7.3.1
rcssmin==1.1.0
redis==4.5.5
requests==2.28.1
rjsmin==1.2.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.2
stack-data==0.6.2
tinycss2==1.1.1
traitlets==5.9.0
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.12
vine==5.0.0
virtualenv==20.17.1
wcwidth==0.2.6
weasyprint==59.0
webencodings==0.5.1
Werkzeug==2.2.2
whitenoise==6.3.0
XlsxWriter==3.0.7
zipp==3.10.0
zopfli==0.2.2
pyotp==2.9.0
twilio==8.8.0
sshtunnel==0.4.0
paramiko==3.3.1
pyshorteners==1.0.1